ca avance petit a petit le nid tout ca tout ca

+++ functions/ft_strlen.c
@@ -9,10 +9,11 @@
 /*   Updated: 2022/10/04 08:11:26 by ldrieske         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(char *a)
+size_t	ft_strlen(char *a)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (a[i])
